+++ R/ggplot_rsi.R
@@ -20,16 +20,17 @@
 #'
 #' Use these functions to create bar plots for antimicrobial resistance analysis. All functions rely on internal \code{\link[ggplot2]{ggplot}} functions.
 #' @param data a \code{data.frame} with column(s) of class \code{"rsi"} (see \code{\link{as.rsi}})
-#' @param position position adjustment of bars, either \code{"stack"} (default) or \code{"dodge"}
+#' @param position position adjustment of bars, either \code{"stack"} (default when \code{fun} is \code{\link{portion_df}}) or \code{"dodge"} (default when \code{fun} is \code{\link{count_df}})
 #' @param x variable to show on x axis, either \code{"Antibiotic"} (default) or \code{"Interpretation"} or a grouping variable
 #' @param fill variable to categorise using the plots legend, either \code{"Antibiotic"} (default) or \code{"Interpretation"} or a grouping variable
 #' @param facet variable to split plots by, either \code{"Interpretation"} (default) or \code{"Antibiotic"} or a grouping variable
 #' @param translate_ab a column name of the \code{\link{antibiotics}} data set to translate the antibiotic abbreviations into, using \code{\link{abname}}. Default behaviour is to translate to official names according to the WHO. Use \code{translate_ab = FALSE} to disable translation.
+#' @param fun function to transform \code{data}, either \code{\link{portion_df}} (default) or \code{\link{count_df}}
 #' @param ... other parameters passed on to \code{\link[ggplot2]{facet_wrap}}
 #' @details At default, the names of antibiotics will be shown on the plots using \code{\link{abname}}. This can be set with the option \code{get_antibiotic_names} (a logical value), so change it e.g. to \code{FALSE} with \code{options(get_antibiotic_names = FALSE)}.
 #'
 #' \strong{The functions}\cr
-#' \code{geom_rsi} will take any variable from the data that has an \code{rsi} class (created with \code{\link{as.rsi}}) using \code{\link{portion_df}} and will plot bars with the percentage R, I and S. The default behaviour is to have the bars stacked and to have the different antibiotics on the x axis.
+#' \code{geom_rsi} will take any variable from the data that has an \code{rsi} class (created with \code{\link{as.rsi}}) using \code{fun} (\code{\link{portion_df}} at default, could also be \code{\link{count_df}}) and will plot bars with the percentage R, I and S. The default behaviour is to have the bars stacked and to have the different antibiotics on the x axis.
 #'
 #' \code{facet_rsi} creates 2d plots (at default based on S/I/R) using \code{\link[ggplot2]{facet_wrap}}.
 #'
@@ -63,6 +64,11 @@
 #' septic_patients %>%
 #'   select(amox, nitr, fosf, trim, cipr) %>%
 #'   ggplot_rsi()
+#'
+#' # get counts instead of percentages:
+#' septic_patients %>%
+#'   select(amox, nitr, fosf, trim, cipr) %>%
+#'   ggplot_rsi(fun = count_df)
 #' \donttest{
 #' # it also supports groups (don't forget to use the group on `x` or `facet`):
 #' septic_patients %>%
@@ -102,25 +108,34 @@
 #'        x = "Microorganisms")
 #' }
 ggplot_rsi <- function(data,
-                       position = "stack",
+                       position = NULL,
                        x = "Antibiotic",
                        fill = "Interpretation",
                        facet = NULL,
                        translate_ab = "official",
+                       fun = portion_df,
                        ...) {
 
   if (!"ggplot2" %in% rownames(installed.packages())) {
     stop('this function requires the ggplot2 package.', call. = FALSE)
   }
 
+  fun_name <- deparse(substitute(fun))
+  if (!fun_name %in% c("portion_df", "count_df")) {
+    stop("`fun` must be portion_df or count_df")
+  }
+
   p <- ggplot2::ggplot(data = data) +
-    geom_rsi(position = position, x = x, fill = fill, translate_ab = translate_ab) +
-    scale_y_percent() +
+    geom_rsi(position = position, x = x, fill = fill, translate_ab = translate_ab, fun = fun) +
     theme_rsi()
 
   if (fill == "Interpretation") {
     # set RSI colours
     p <- p + scale_rsi_colours()
+  }
+  if (fun_name == "portion_df") {
+    # portions, so use y scale with percentage
+    p <- p + scale_y_percent()
   }
 
   if (!is.null(facet)) {
@@ -132,10 +147,27 @@ ggplot_rsi <- function(data,
 
 #' @rdname ggplot_rsi
 #' @export
-geom_rsi <- function(position = "stack",
+geom_rsi <- function(position = NULL,
                      x = c("Antibiotic", "Interpretation"),
                      fill = "Interpretation",
-                     translate_ab = "official")  {
+                     translate_ab = "official",
+                     fun = portion_df)  {
+
+  fun_name <- deparse(substitute(fun))
+  if (!fun_name %in% c("portion_df", "count_df", "fun")) {
+    stop("`fun` must be portion_df or count_df")
+  }
+  if (identical(fun, count_df)) {
+    y <- "Count"
+    if (missing(position) | is.null(position)) {
+      position <- "dodge"
+    }
+  } else {
+    y <- "Percentage"
+    if (missing(position) | is.null(position)) {
+      position <- "stack"
+    }
+  }
 
   x <- x[1]
   if (x %in% tolower(c('ab', 'antibiotic', 'abx', 'antibiotics'))) {
@@ -147,8 +179,8 @@ geom_rsi <- function(position = "stack",
   options(get_antibiotic_names = translate_ab)
 
   ggplot2::layer(geom = "bar", stat = "identity", position = position,
-                 mapping = ggplot2::aes_string(x = x, y = "Percentage", fill = fill),
-                 data = AMR::portion_df, params = list())
+                 mapping = ggplot2::aes_string(x = x, y = y, fill = fill),
+                 data = fun, params = list())
 
 }
 
@@ -163,14 +195,13 @@ facet_rsi <- function(facet = c("Interpretation", "Antibiotic"), ...) {
     facet <- "Antibiotic"
   }
 
-  ggplot2::facet_wrap(facets = facet, scales = "free", ...)
+  ggplot2::facet_wrap(facets = facet, scales = "free_x", ...)
 }
 
 #' @rdname ggplot_rsi
 #' @export
 scale_y_percent <- function() {
-  ggplot2::scale_y_continuous(name = "Percentage",
-                              breaks = seq(0, 1, 0.1),
+  ggplot2::scale_y_continuous(breaks = seq(0, 1, 0.1),
                               limits = c(0, 1),
                               labels = percent(seq(0, 1, 0.1)))
 }
@@ -184,8 +215,8 @@ scale_rsi_colours <- function() {
 #' @rdname ggplot_rsi
 #' @export
 theme_rsi <- function() {
-  theme_minimal() +
-    theme(panel.grid.major.x = element_blank(),
-          panel.grid.minor = element_blank(),
-          panel.grid.major.y = element_line(colour = "grey75"))
+  ggplot2::theme_minimal() +
+    ggplot2::theme(panel.grid.major.x = ggplot2::element_blank(),
+                   panel.grid.minor = ggplot2::element_blank(),
+                   panel.grid.major.y = ggplot2::element_line(colour = "grey75"))
 }